node: move [ParseCiliumNode] to the k8s package
Move the [node.ParseCiliumNode] function out of the node/types package,
so that we don't need to import a significant amount of Kubernetes-related
packages from a simple types package.
